tools350/Im2MIF/requirements.txt
typing==3.6.4
numpy==1.14.3
Pillow==6.2.0
scikit_learn==0.20.3
